+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [4.2.0] - 2021-08-09
+
 ### Added
 
 - Support initial user password (only set doing user creation). See [docs/FEATURE.md](docs/FEATURES.md) for more information.
@@ -465,7 +467,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 [3.4.0-rc2]: https://github.com/adorsys/keycloak-config-cli/compare/v3.4.0-rc1...v3.4.0-rc2
 [3.4.0-rc1]: https://github.com/adorsys/keycloak-config-cli/compare/v3.4.0-rc0...v3.4.0-rc1
 [3.4.0-rc0]: https://github.com/adorsys/keycloak-config-cli/compare/v3.3.1...v3.4.0-rc0
-[Unreleased]: https://github.com/adorsys/keycloak-config-cli/compare/v4.1.0...HEAD
+[Unreleased]: https://github.com/adorsys/keycloak-config-cli/compare/v4.2.0...HEAD
+[4.2.0]: https://github.com/adorsys/keycloak-config-cli/compare/v4.1.0...v4.2.0
 [4.1.0]: https://github.com/adorsys/keycloak-config-cli/compare/v4.0.1...v4.1.0
 [4.0.1]: https://github.com/adorsys/keycloak-config-cli/compare/v4.0.0...v4.0.1
 [4.0.0]: https://github.com/adorsys/keycloak-config-cli/compare/vBreaking...v4.0.0